refactor: rename components, move layouts to root
