+++ src/needs-refactoring/pages/[placeType]/[id]/report/mail-to-support.tsx
@@ -2,8 +2,8 @@ import { Button } from "@radix-ui/themes";
 import type { LocalizedString } from "@sozialhelden/a11yjson";
 import { t } from "@transifex/native";
 import React, { type FC, useContext } from "react";
-import { getLayout } from "~/components/layouts/DefaultLayout";
 import { useUserAgent } from "~/hooks/useUserAgent";
+import { getLayout } from "~/layouts/DefaultLayout";
 import { AppStateLink } from "~/needs-refactoring/components/App/AppStateLink";
 import { FeaturePanelContext } from "~/needs-refactoring/components/CombinedFeaturePanel/FeaturePanelContext";
 import { StyledReportView } from "~/needs-refactoring/components/CombinedFeaturePanel/ReportView";
