+++ src/needs-refactoring/pages/events/index.tsx
@@ -1,5 +1,5 @@
+import { getLayout } from "~/layouts/DefaultLayout";
 import MappingEventListPanel from "~/needs-refactoring/components/MappingEvents/MappingEventListPanel";
-import { getLayout } from "~/components/layouts/DefaultLayout";
 
 export default function Page() {
   return <MappingEventListPanel />;
