+++ src/pages/[placeType]/[id]/images/[imageId]/index.tsx
@@ -1,8 +1,8 @@
 import { useRouter } from "next/router";
 import { useContext, useMemo } from "react";
+import { getLayout } from "~/layouts/FeatureDetailsLayout";
 import { CombinedFeaturePanel } from "~/needs-refactoring/components/CombinedFeaturePanel/CombinedFeaturePanel";
 import { FeaturePanelContext } from "~/needs-refactoring/components/CombinedFeaturePanel/FeaturePanelContext";
-import { getLayout } from "~/components/layouts/FeatureDetailsLayout";
 
 export default function ShowImagePage() {
   const { features } = useContext(FeaturePanelContext);
